:bug: bugfix: 버그 수정 (#30)
- 데이터 조회시 where절에 storeId 추가
