+++ src/components/order/success/SuccessContainer.tsx
@@ -4,8 +4,8 @@ import useOrderStore, { addOrderId, getTotalPrice } from '@/shared/store/order';
 import { groupByKey } from '@/shared/helper';
 import { Tables } from '@/types/supabase';
 import { useStoreQuery } from '@/hooks/store/useStoreQuery';
-import { useStoreOrderQuery } from '@/hooks/order/useStoreOrderQuery';
-import { useNumberOrderQuery } from '@/hooks/order/useNumberOrderQuery';
+import { useStoreOrderSetQuery } from '@/hooks/order/useStoreOrderSetQuery';
+import { useNumberOrderSetQuery } from '@/hooks/order/useNumberOrderSetQuery';
 import styles from './styles/SuccessContainer.module.css';
 import Image from 'next/image';
 import image from '@/../public/images/image-success.png';
@@ -20,8 +20,8 @@ const SuccessContainer = ({ payment }: { payment?: Payment }) => {
   const orderType = useOrderStore(state => state.orderType);
 
   const { addSales } = useSalesQuery();
-  const { addStoreOrder } = useStoreOrderQuery();
-  const { addNumberOrder } = useNumberOrderQuery();
+  const { addStoreOrder } = useStoreOrderSetQuery();
+  const { addNumberOrder } = useNumberOrderSetQuery();
   const { incrementOrderNumber } = useStoreQuery();
   const [isPageLoading, setIsPageLoading] = useState(false);
   const router = useRouter();
