+++ src/server/api/supabase/order-togo.ts
@@ -6,10 +6,11 @@ export const addNumberOrder = async (item: Omit<Tables<'order_number'>, 'id'>) =
   return { data, error };
 };
 
-export const fetchNumberOrderByOrderIdWithStoreName = async (orderIdList: string[]) => {
+export const fetchNumberOrderByOrderIdWithStoreName = async (orderIdList: string[], storeId: string) => {
   const { data, error } = await supabase
     .from('order_number')
     .select('*, store(business_name)')
+    .eq('store_id', storeId)
     .filter('order_id', 'in', `(${orderIdList.join(',')})`);
   return { data, error };
 };
